# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API App - Clinica",
        "version": "1.0"
    },
    "paths": {
        "/api/show-history/cita-control/{user}": {
            "get": {
                "tags": [
                    "cita-control"
                ],
                "summary": "Mostrando historial de cita de control del cliente",
                "operationId": "showHistorialCitaControlUser",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "Id of User",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Showing history of citas of control of the Client"
                    },
                    "401": {
                        "description": "User not authenticated"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Unexpected error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/show/cita-control/{user}": {
            "get": {
                "tags": [
                    "cita-control"
                ],
                "summary": "Mostrando cita de control del cliente",
                "operationId": "showCitaControlUser",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "Id of User",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Showing cita of control of the Client"
                    },
                    "401": {
                        "description": "User not authenticated"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Unexpected error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/create/cita-control": {
            "post": {
                "tags": [
                    "cita-control"
                ],
                "summary": "Crear cita de control",
                "operationId": "createCitaControl",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "peso": {
                                        "type": "integer"
                                    },
                                    "musculo": {
                                        "type": "integer"
                                    },
                                    "grasas": {
                                        "type": "integer"
                                    },
                                    "porcentaje_grasa": {
                                        "type": "integer"
                                    },
                                    "cc": {
                                        "type": "integer"
                                    },
                                    "grasa_viceral": {
                                        "type": "integer"
                                    },
                                    "evolucion": {
                                        "type": "string"
                                    },
                                    "cliente_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cita de control created"
                    },
                    "422": {
                        "description": "Validation rules failed"
                    },
                    "500": {
                        "description": "Unexpected error"
                    }
                }
            }
        },
        "/api/update/cita-control/{cita}": {
            "post": {
                "tags": [
                    "cita-control"
                ],
                "summary": "Actualizar cita de control",
                "operationId": "updateCitaControl",
                "parameters": [
                    {
                        "name": "cita",
                        "in": "path",
                        "description": "Id of the Cita de Control",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "peso": {
                                        "type": "integer"
                                    },
                                    "musculo": {
                                        "type": "integer"
                                    },
                                    "grasas": {
                                        "type": "integer"
                                    },
                                    "porcentaje_grasa": {
                                        "type": "integer"
                                    },
                                    "cc": {
                                        "type": "integer"
                                    },
                                    "grasa_viceral": {
                                        "type": "integer"
                                    },
                                    "evolucion": {
                                        "type": "string"
                                    },
                                    "cliente_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cita de control updated"
                    },
                    "422": {
                        "description": "Validation rules failed"
                    },
                    "500": {
                        "description": "Unexpected error"
                    }
                }
            }
        },
        "/api/eliminar/cita-control/{cita}": {
            "delete": {
                "tags": [
                    "cita-control"
                ],
                "summary": "Delete cita de control",
                "operationId": "deleteCitaControl",
                "parameters": [
                    {
                        "name": "cita",
                        "in": "path",
                        "description": "Id of the cita control",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cita control deleted succesfully"
                    },
                    "404": {
                        "description": "Cita control not found"
                    },
                    "500": {
                        "description": "Unexpected error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    }
}